+++ components/Button/Bookmark/Subscribe.tsx
@@ -23,7 +23,7 @@ const Subscribe = ({
   disabled
 }: {
   article: BookmarkArticle
-  size: 'small' | 'default'
+  size: 'xsmall' | 'small' | 'default'
   disabled?: boolean
 }) => (
   <Mutation
@@ -47,14 +47,14 @@ const Subscribe = ({
         <Icon
           size={size}
           id={
-            size === 'small'
-              ? ICON_BOOKMARK_SM_INACTIVE.id
-              : ICON_BOOKMARK_REGULAR_INACTIVE.id
+            size === 'default'
+              ? ICON_BOOKMARK_REGULAR_INACTIVE.id
+              : ICON_BOOKMARK_SM_INACTIVE.id
           }
           viewBox={
-            size === 'small'
-              ? ICON_BOOKMARK_SM_INACTIVE.viewBox
-              : ICON_BOOKMARK_REGULAR_INACTIVE.viewBox
+            size === 'default'
+              ? ICON_BOOKMARK_REGULAR_INACTIVE.viewBox
+              : ICON_BOOKMARK_SM_INACTIVE.viewBox
           }
         />
       </button>