+++ components/Placeholder/utils.ts
@@ -1,5 +1,5 @@
 export const LoaderProps = {
   ariaLabel: '加載中',
-  primaryColor: '#dddddd',
-  secondaryColor: '#f6f6f6'
+  primaryColor: '#f6f6f6',
+  secondaryColor: '#dddddd'
 }