+++ components/UserProfile/index.tsx
@@ -1,7 +1,6 @@
 import classNames from 'classnames'
 import gql from 'graphql-tag'
-import _get from 'lodash/get'
-import _some from 'lodash/some'
+import { get, some } from 'lodash'
 import Link from 'next/link'
 import { withRouter, WithRouterProps } from 'next/router'
 import { useContext, useState } from 'react'
@@ -11,6 +10,8 @@ import { Avatar, Placeholder, Tooltip, Translate } from '~/components'
 import { FollowButton } from '~/components/Button/Follow'
 import { Query } from '~/components/GQL'
 import { Icon } from '~/components/Icon'
+import ShareButton from '~/components/ShareButton'
+import ShareModal from '~/components/ShareButton/ShareModal'
 import { TextIcon } from '~/components/TextIcon'
 import { UserProfileEditor } from '~/components/UserProfileEditor'
 import { ViewerContext } from '~/components/Viewer'
@@ -21,6 +22,8 @@ import ICON_SEED_BADGE from '~/static/icons/eerly-user-badge.svg?sprite'
 import ICON_LOCK from '~/static/icons/lock.svg?sprite'
 
 import { UserProfileUser } from './__generated__/UserProfileUser'
+import Cover from './Cover'
+import Description from './Description'
 import EditProfileButton from './EditProfileButton'
 import styles from './styles.css'
 
@@ -35,6 +38,7 @@ const fragments = {
           type
         }
         description
+        profileCover
       }
       followees(input: { first: 0 }) {
         totalCount
@@ -77,7 +81,7 @@ const SeedBadge = () => (
       <Icon
         id={ICON_SEED_BADGE.id}
         viewBox={ICON_SEED_BADGE.viewBox}
-        style={{ width: 16, height: 16, marginLeft: '0.5rem' }}
+        style={{ width: 16, height: 16 }}
       />
     </span>
   </Tooltip>
@@ -91,7 +95,7 @@ const OnboardingBadge = () => (
         <Icon
           id={ICON_LOCK.id}
           viewBox={ICON_LOCK.viewBox}
-          style={{ width: 16, height: 16, marginLeft: '0.5rem' }}
+          style={{ width: 16, height: 16 }}
         />
       }
     >
@@ -99,6 +103,17 @@ const OnboardingBadge = () => (
     </TextIcon>
   </span>
 )
+
+const CoverContainer = ({ children }: any) => (
+  <>
+    <div className="cover-container l-row">
+      <section className="l-col-4 l-col-md-8 l-col-lg-12">{children}</section>
+    </div>
+    <style jsx>{styles}</style>
+  </>
+)
+
+type UserProfileResultType = QueryResult & { data: UserProfileUser }
 
 const BaseUserProfile: React.FC<WithRouterProps> = ({ router }) => {
   const viewer = useContext(ViewerContext)
@@ -114,52 +129,62 @@ const BaseUserProfile: React.FC<WithRouterProps> = ({ router }) => {
 
   return (
     <section className={containerClass}>
-      <div className="content-container l-row">
-        <section className="l-col-4 l-col-md-6 l-offset-md-1 l-col-lg-8 l-offset-lg-2">
-          <Query
-            query={isMe ? ME_PROFILE : USER_PROFILE}
-            variables={isMe ? {} : { userName }}
-          >
-            {({
-              data,
-              loading,
-              error
-            }: QueryResult & { data: UserProfileUser }) => {
-              if (loading) {
-                return <Placeholder.UserProfile />
-              }
+      <Query
+        query={isMe ? ME_PROFILE : USER_PROFILE}
+        variables={isMe ? {} : { userName }}
+      >
+        {({ data, loading, error }: UserProfileResultType) => {
+          if (loading) {
+            return (
+              <CoverContainer>
+                <Placeholder.UserProfile />
+              </CoverContainer>
+            )
+          }
 
-              if (isMe && editing) {
-                return (
-                  <UserProfileEditor
-                    user={data.viewer}
-                    saveCallback={setEditing}
-                  />
-                )
-              }
+          if (isMe && editing) {
+            return (
+              <UserProfileEditor user={data.viewer} saveCallback={setEditing} />
+            )
+          }
 
-              const user = isMe ? data.viewer : data.user
-              const userFollowersPath = toPath({
-                page: 'userFollowers',
-                userName: user.userName
-              })
-              const userFolloweesPath = toPath({
-                page: 'userFollowees',
-                userName: user.userName
-              })
-              const badges = _get(user, 'info.badges', [])
-              const hasSeedBadge = _some(badges, { type: 'seed' })
-              const state = _get(user, 'status.state')
-              const isOnboarding = state === 'onboarding'
+          const user = isMe ? data.viewer : data.user
+          const userFollowersPath = toPath({
+            page: 'userFollowers',
+            userName: user.userName
+          })
+          const userFolloweesPath = toPath({
+            page: 'userFollowees',
+            userName: user.userName
+          })
+          const badges = get(user, 'info.badges', [])
+          const hasSeedBadge = some(badges, { type: 'seed' })
+          const state = get(user, 'status.state')
+          const isOnboarding = state === 'onboarding'
+          const profileCover = get(user, 'info.profileCover', '')
 
-              // me profile
-              if (isMe) {
-                return (
+          return (
+            <>
+              <CoverContainer>
+                <Cover cover={profileCover} />
+              </CoverContainer>
+              <div className="content-container l-row">
+                <section className="l-col-4 l-col-md-6 l-offset-md-1 l-col-lg-8 l-offset-lg-2">
                   <section className="content">
-                    <Avatar
-                      size="xlarge"
-                      user={viewer.isInactive ? undefined : user}
-                    />
+                    <div className="avatar-container">
+                      <Avatar
+                        size="xlarge"
+                        user={!isMe && viewer.isInactive ? undefined : user}
+                      />
+                      {!isMe && (
+                        <section className="action-button">
+                          <FollowButton user={user} size="default" />
+                          <span className="u-sm-up-hide">
+                            <ShareButton />
+                          </span>
+                        </section>
+                      )}
+                    </div>
 
                     <section className="info">
                       <header className="header">
@@ -169,6 +194,7 @@ const BaseUserProfile: React.FC<WithRouterProps> = ({ router }) => {
                               {user.displayName}
                               {hasSeedBadge && <SeedBadge />}
                               {isOnboarding && <OnboardingBadge />}
+                              {!isMe && <FollowButton.State user={user} />}
                             </span>
                           )}
                           {viewer.isArchived && (
@@ -197,16 +223,18 @@ const BaseUserProfile: React.FC<WithRouterProps> = ({ router }) => {
                           )}
                         </section>
                         <section className="action-button">
-                          {!viewer.isInactive && (
+                          {isMe && !viewer.isInactive && (
                             <EditProfileButton setEditing={setEditing} />
                           )}
+                          <span className={!isMe ? 'u-sm-down-hide' : ''}>
+                            <ShareButton />
+                          </span>
                         </section>
                       </header>
 
                       {!viewer.isInactive && (
-                        <p className="description">{user.info.description}</p>
+                        <Description description={user.info.description} />
                       )}
-
                       <section className="info-follow">
                         <Link {...userFollowersPath}>
                           <a className="followers">
@@ -214,8 +242,8 @@ const BaseUserProfile: React.FC<WithRouterProps> = ({ router }) => {
                               {numAbbr(user.followers.totalCount)}
                             </span>
                             <Translate
-                              zh_hant={TEXT.zh_hant.followingMe}
-                              zh_hans={TEXT.zh_hans.followingMe}
+                              zh_hant={TEXT.zh_hant.follower}
+                              zh_hans={TEXT.zh_hans.follower}
                             />
                           </a>
                         </Link>
@@ -225,71 +253,22 @@ const BaseUserProfile: React.FC<WithRouterProps> = ({ router }) => {
                               {numAbbr(user.followees.totalCount)}
                             </span>
                             <Translate
-                              zh_hant={TEXT.zh_hant.myFollowees}
-                              zh_hans={TEXT.zh_hans.myFollowees}
+                              zh_hant={TEXT.zh_hant.following}
+                              zh_hans={TEXT.zh_hans.following}
                             />
                           </a>
                         </Link>
                       </section>
                     </section>
                   </section>
-                )
-              }
-
-              // other's profile
-              return (
-                <section className="content">
-                  <Avatar size="xlarge" user={user} />
-
-                  <section className="info">
-                    <header className="header">
-                      <section className="name">
-                        <span>
-                          {user.displayName}
-                          {hasSeedBadge && <SeedBadge />}
-                          {isOnboarding && <OnboardingBadge />}
-                        </span>
-                        <FollowButton.State user={user} />
-                      </section>
-
-                      <section className="action-button">
-                        <FollowButton user={user} size="default" />
-                      </section>
-                    </header>
-
-                    <p className="description">{user.info.description}</p>
-
-                    <section className="info-follow">
-                      <Link {...userFollowersPath}>
-                        <a className="followers">
-                          <span className="count">
-                            {numAbbr(user.followers.totalCount)}
-                          </span>
-                          <Translate
-                            zh_hant={TEXT.zh_hant.follower}
-                            zh_hans={TEXT.zh_hans.follower}
-                          />
-                        </a>
-                      </Link>
-                      <Link {...userFolloweesPath}>
-                        <a className="followees">
-                          <span className="count">
-                            {numAbbr(user.followees.totalCount)}
-                          </span>
-                          <Translate
-                            zh_hant={TEXT.zh_hant.following}
-                            zh_hans={TEXT.zh_hans.following}
-                          />
-                        </a>
-                      </Link>
-                    </section>
-                  </section>
                 </section>
-              )
-            }}
-          </Query>
-        </section>
-      </div>
+              </div>
+            </>
+          )
+        }}
+      </Query>
+
+      <ShareModal />
 
       <style jsx>{styles}</style>
     </section>