+++ views/User/Articles/UserArticles.tsx
@@ -1,19 +1,26 @@
 import gql from 'graphql-tag'
 import _get from 'lodash/get'
 import { withRouter, WithRouterProps } from 'next/router'
-import { useContext } from 'react'
 import { QueryResult } from 'react-apollo'
 
-import { ArticleDigest, Head, InfiniteScroll, Placeholder } from '~/components'
+import {
+  ArticleDigest,
+  Head,
+  Icon,
+  InfiniteScroll,
+  Placeholder
+} from '~/components'
 import EmptyArticle from '~/components/Empty/EmptyArticle'
 import { Query } from '~/components/GQL'
-import { ViewerContext } from '~/components/Viewer'
+import { Translate } from '~/components/Language'
 
 import { ANALYTICS_EVENTS, FEED_TYPE } from '~/common/enums'
 import { analytics, getQuery, mergeConnections } from '~/common/utils'
 import ICON_192 from '~/static/icon-192x192.png?url'
+import ICON_DOT_DIVIDER from '~/static/icons/dot-divider.svg?sprite'
 
 import { UserArticleFeed } from './__generated__/UserArticleFeed'
+import styles from './styles.css'
 
 const USER_ARTICLES_FEED = gql`
   query UserArticleFeed(
@@ -28,8 +35,11 @@ const USER_ARTICLES_FEED = gql`
       displayName
       info {
         description
+        totalWordCount
+        profileCover
       }
       articles(input: { first: 10, after: $cursor }) {
+        totalCount
         pageInfo {
           startCursor
           endCursor
@@ -47,11 +57,35 @@ const USER_ARTICLES_FEED = gql`
   ${ArticleDigest.Feed.fragments.article}
 `
 
+const ArticleSummaryInfo = ({ data }: { data: UserArticleFeed }) => {
+  const { totalWordCount: words } = _get(data, 'user.info', {
+    totalWordCount: 0
+  })
+  const { totalCount: articles } = _get(data, 'user.articles', {
+    totalCount: 0
+  })
+  return (
+    <>
+      <div className="info">
+        <Translate zh_hant="創作了" zh_hans="创作了" />
+        <span>{articles}</span>
+        <Translate zh_hant="篇作品" zh_hans="篇作品" />
+        <Icon
+          id={ICON_DOT_DIVIDER.id}
+          viewBox={ICON_DOT_DIVIDER.viewBox}
+          style={{ width: 18, height: 18 }}
+        />
+        <Translate zh_hant="累積創作" zh_hans="累积创作" />
+        <span>{words}</span>
+        <Translate zh_hant="字" zh_hans="字" />
+      </div>
+      <style jsx>{styles}</style>
+    </>
+  )
+}
+
 const UserArticles: React.FC<WithRouterProps> = ({ router }) => {
   const userName = getQuery({ router, key: 'userName' })
-
-  const viewer = useContext(ViewerContext)
-  const isMe = viewer.userName === userName
 
   if (!userName) {
     return null
@@ -96,7 +130,7 @@ const UserArticles: React.FC<WithRouterProps> = ({ router }) => {
               zh_hans: `${data.user.displayName}的创作空间站`
             }}
             description={data.user.info.description}
-            image={ICON_192}
+            image={data.user.info.profileCover || ICON_192}
           />
         )
 
@@ -112,6 +146,7 @@ const UserArticles: React.FC<WithRouterProps> = ({ router }) => {
         return (
           <>
             <CustomHead />
+            <ArticleSummaryInfo data={data} />
             <InfiniteScroll
               hasNextPage={pageInfo.hasNextPage}
               loadMore={loadMore}
@@ -132,7 +167,7 @@ const UserArticles: React.FC<WithRouterProps> = ({ router }) => {
                         article={node}
                         hasBookmark
                         hasDateTime
-                        hasFingerprint={isMe}
+                        hasFingerprint
                         hasState
                       />
                     </li>